ibc: ⛅ hoist ics02 validation out of client_counter
nb: this commit strictly contains plain code motion. it does not make any
changes to the code being moved.

`ics02_validation` contains some important code for interfacing with
tendermint/cometbft. while this submodule lives beneath the client
counter logic, its contents are used in our message handling code
related to e.g. misbehavior.

this hoists this submodule up one level, to be a child of
`penumbra_ibc::component`.
