+++ crates/core/component/ibc/src/component/msg_handler/update_client.rs
@@ -18,8 +18,7 @@ use tendermint_light_client_verifier::{
 
 use crate::component::{
     client::{Ics2ClientExt as _, StateReadExt as _, StateWriteExt as _},
-    client_counter::ics02_validation,
-    MsgHandler,
+    ics02_validation, MsgHandler,
 };
 
 #[async_trait]
